# worker/openapi.yaml
openapi: 3.0.3
info:
  title: Acervo Filarmonica API
  description: API do sistema de gestão de partituras da Filarmonica 25 de Março
  version: 1.0.0
  contact:
    name: Acervo Filarmonica
servers:
  - url: http://localhost:8787
    description: Desenvolvimento local
  - url: https://acervo-filarmonica-api.acssjr.workers.dev
    description: Produção
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
paths:
...
  /api/ensaios/{data}/partituras:
    get:
      tags: [Repertorio]
      operationId: getPartiturasEnsaio
      parameters:
        - { name: data, in: path, required: true, schema: { type: string, format: date } }
      responses:
        '200': { description: Partituras do ensaio, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '400': { description: Data inválida, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      tags: [Repertorio]
      operationId: addPartituraEnsaio
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: data, in: path, required: true, schema: { type: string, format: date } }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [partitura_id], properties: { partitura_id: { type: integer } } } } }
      responses:
        '200': { description: Partitura adicionada, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
  /api/ensaios/{data}/partituras/{partituraId}:
    delete:
      tags: [Repertorio]
      operationId: deletePartituraEnsaio
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: data, in: path, required: true, schema: { type: string, format: date } }
        - { name: partituraId, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Partitura removida, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
  /api/ensaios/{data}/partituras/reorder:
    put:
      tags: [Repertorio]
      operationId: reorderPartiturasEnsaio
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: data, in: path, required: true, schema: { type: string, format: date } }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: Ordem atualizada, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '400': { description: Ordem inválida, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/ensaios/{data}/config:
    patch:
      tags: [Repertorio]
      operationId: updateEnsaioConfig
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: data, in: path, required: true, schema: { type: string, format: date } }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, properties: { youtube_url: { type: string, format: uri, nullable: true } } } } }
      responses:
        '200': { description: Configuração atualizada, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '400': { description: Data ou URL inválida, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }